moved UtilsTests to proper directory
