+++ utilities/utilities.py
@@ -499,16 +499,3 @@ def mask_quoted_text(text: str) -> str:
     for start, end in quote_inds:
         text = text[:start] + (end - start) * "\ufeff" + text[end:]
     return text
-
-
-class UtilsTests:
-    def test_split_message_for_discord(self):
-        test_out = len(
-            Utilities.split_message_for_discord(
-                "123456789012345\n1234567890123456789\n10\n10\n10\n01234567890123456789",
-                max_length=20,
-            )
-        )
-        self.assertEqual(len(test_out), 4)
-        for chunk in test_out:
-            self.assertLessEqual(len(chunk), 20)
